[SW-225][added] 리베이스 후 충돌처리
